refactor: 브랜드 감지 match 함수를 cardBrandMatchers.ts로 분리
